homestretch!!
Took everything wrong with the old code, took all the unnecessary clutter and the loads of duplicates out, and reworked teh base GUI based on what Dr. Kancharla did. After 5  or 6 hours of nonstsop trial-and-error and nonstop debugging on solely the rasberry pi (which is why there were no commits before ths), I was able to rework the game and ACTUALLY MAKE IT WORK!!!!!! YAYYYYYYY